editing from-stdin.fish

+++ from-stdin.fish
@@ -1,10 +1,14 @@
 # Defined in /home/michael/.config/fish/buffer/from-stdin.fish @ line 2
 function from-stdin
-	set tmp /tmp/stdin-(uid)
-  rm $tmp
-  while read -l line
-      echo $line >> $tmp
-  end
-  emacsclient -nw $tmp
-  cat $tmp
+	if test -f $tmp
+        rm $tmp
+    end
+    set tmp /tmp/stdin-(uid)
+    rm $tmp
+    while read -l line
+        echo $line >> $tmp
+    end
+    emacsclient -nw $tmp
+    cat $tmp
+    rm $tmp
 end
